fix: change /cogs for /commands for a better structure
fix: the words were separated into another file
