+++ src/commands/avatar.js
