+++ src/commands/hola/palabras.js
@@ -1,0 +1,19 @@
+const palabras = [
+  'Hola, que tal',
+  'Hi, how are you?',
+  'hola, como estas',
+  'Que tal',
+  'Mucho gusto',
+  'Que haces',
+  'Mucho tiempo sin verte',
+  'Cómo está usted',
+  'Que gusto verte',
+  '¿Cómo van las cosas?',
+  '¿Qué pasa?',
+  '¿Qué hay de nuevo?',
+  'Buenas',
+  'Hello world',
+  'Hola mundo',
+]
+
+module.exports = palabras
